refactor: move APIKeyGenerator and KeyGenerator to client package
