+++ profanity-domain/src/main/java/app/application/client/APIKeyGenerator.java
@@ -1,4 +1,4 @@
-package app.application.apikey;
+package app.application.client;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
